Переносит вспомогательные функции в utils из form
