+++ js/form.js
@@ -1,3 +1,5 @@
+import { validateHastags } from './utils.js';
+
 const body = document.querySelector('body');
 const overlay = body.querySelector('.img-upload__overlay');
 const form = body.querySelector('.img-upload__form');
@@ -5,10 +7,6 @@ const fileField = form.querySelector('#upload-file');
 const cancelButton = form.querySelector('.img-upload__cancel');
 const hashtagField = form.querySelector('.text__hashtags');
 const commentField = form.querySelector('.text__description');
-
-const MAX_COUNT_TAGS = 5;
-const TAGS_VALID = /^#[a-zа-яё0-9]{1,19}$/i;
-
 
 const showModal = () => {
   overlay.classList.remove('hidden');
@@ -36,20 +34,6 @@ const pristine = new Pristine(form, {
   errorTextParent: 'img-upload__field-wrapper',
   errorClass: 'img-upload__field-wrapper__error-message'
 });
-
-const getTagList = (tags) => {
-  const uniqueHashtags = new Set(tags.toLowerCase().split(' ').filter((tag) => tag.trim()));
-  return Array.from(uniqueHashtags);
-};
-const isValidCount = (tagList) => tagList.length <= MAX_COUNT_TAGS;
-
-const isValidHastag = (hashtag) => TAGS_VALID.test(hashtag);
-
-const validateHastags = (tags) => {
-  const tagList = getTagList(tags);
-
-  return isValidCount(tagList) && tagList.every(isValidHastag);
-};
 
 pristine.addValidator(
   hashtagField,
