+++ js/utils.js
@@ -1,3 +1,20 @@
+const MAX_COUNT_TAGS = 5;
+const TAGS_VALID = /^#[a-zа-яё0-9]{1,19}$/i;
+
+const getTagList = (tags) => {
+  const uniqueHashtags = new Set(tags.toLowerCase().split(' ').filter((tag) => tag.trim()));
+  return Array.from(uniqueHashtags);
+};
+const isValidCount = (tagList) => tagList.length <= MAX_COUNT_TAGS;
+
+const isValidHastag = (hashtag) => TAGS_VALID.test(hashtag);
+
+const validateHastags = (tags) => {
+  const tagList = getTagList(tags);
+
+  return isValidCount(tagList) && tagList.every(isValidHastag);
+};
+
 const getRandomInt = (min, max) => {
   if ((max < 0) || (min < 0)) {
     return NaN;
@@ -7,10 +24,9 @@ const getRandomInt = (min, max) => {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
 
-
 const getCorrectLength = (text, maxLenght) => text.length <= maxLenght;
 getCorrectLength('');
 
 const getRandomElementArray = (array) => array[getRandomInt(0, array.length - 1)];
 
-export {getRandomInt, getRandomElementArray};
+export {getRandomInt, getRandomElementArray, validateHastags};
